form-service/requirements.txt
Django>=4.2
djangorestframework>=3.14
psycopg2-binary>=2.9
pytest-django>=4.5
visine-common
dj-database-url>=2.1.0
